Split action controllers into separate package
brokkr-users-5 WIP
