+++ composer.json
@@ -12,15 +12,15 @@
   "minimum-stability": "dev",
   "prefer-stable": true,
   "require": {
-    "php": "^7.3",
-    "container-interop/container-interop": "^1.2",
-    "doctrine/doctrine-orm-module": "^3.0",
+    "php": "7.3.* || 7.4.* || 8.0.*",
+    "doctrine/doctrine-orm-module": "^3.0 || ^4.0",
     "laminas/laminas-authentication": "^2.7",
     "laminas/laminas-component-installer": "^2.0 || ^1.0 || ^0.7",
     "laminas/laminas-mail": "^2.10",
-    "laminas/laminas-mvc": "^3.1",
+    "laminas/laminas-permissions-acl": "^2.8",
+    "laminas/laminas-router": "^3.3",
     "laminas/laminas-servicemanager": "^3.4",
-    "ramsey/uuid": "^4.0",
+    "psr/container": "^1.1",
     "riddlestone/brokkr-acl": "^0.2.0",
     "riddlestone/brokkr-mail": "dev-master"
   },
@@ -33,17 +33,19 @@
     "laminas/laminas-mvc-plugin-flashmessenger": "To provide feedback on successful login, logout, password reset, etc."
   },
   "require-dev": {
+    "doctrine/dbal": "^2.13",
+    "container-interop/container-interop": "^1.2",
+    "laminas/laminas-db": "^2.12",
+    "laminas/laminas-mvc": "^3.2",
     "laminas/laminas-mvc-plugin-flashmessenger": "^1.2",
-    "laminas/laminas-router": "^3.3",
     "laminas/laminas-session": "^2.9",
-    "laminas/laminas-validator": "^2.13",
     "phpmd/phpmd": "^2.8",
     "phpmetrics/phpmetrics": "^2.7",
     "phpunit/phpunit": "^9.4",
-    "sensiolabs-de/deptrac-shim": "^0.8.0",
+    "ramsey/uuid": "^4.0",
+    "qossmic/deptrac-shim": "^0.13.0",
     "squizlabs/php_codesniffer": "^3.5",
-    "vimeo/psalm": "^4.1",
-    "laminas/laminas-db": "^2.12"
+    "vimeo/psalm": "^4.1"
   },
   "autoload-dev": {
     "psr-4": {
