+++ config/module.config.php
@@ -7,11 +7,6 @@ use Laminas\Authentication\AuthenticationService;
 use Laminas\ServiceManager\Factory\InvokableFactory;
 
 return [
-    'acl_resource_manager' => [
-        'abstract_factories' => [
-            Acl\ResourceFactory::class,
-        ],
-    ],
     'acl_role_manager' => [
         'abstract_factories' => [
             Acl\RoleFactory::class,
@@ -23,20 +18,6 @@ return [
         ],
         'providers' => [
             Acl\RoleRelationshipFactory::class,
-        ],
-    ],
-    'acl_rule_manager' => [
-        'factories' => [
-            Acl\RuleProvider::class => InvokableFactory::class,
-        ],
-        'providers' => [
-            Acl\RuleProvider::class,
-        ],
-    ],
-    'controllers' => [
-        'factories' => [
-            Controller\AccountController::class => Controller\AccountControllerFactory::class,
-            Controller\UsersController::class => Controller\UsersControllerFactory::class,
         ],
     ],
     'doctrine' => [
@@ -56,43 +37,6 @@ return [
         ],
     ],
     'global_salt' => 'PLEASE CHANGE ME',
-    'navigation' => [
-        'admin' => [
-            'admin' => [
-                'pages' => [
-                    'users' => [
-                        'label' => 'User Management',
-                        'route' => 'admin/users',
-                        'resource' => Controller\UsersController::class . '::indexAction',
-                    ],
-                ],
-            ],
-        ],
-        'personal' => [
-            'admin' => [
-                'pages' => [
-                    'users' => [
-                        'label' => 'User Management',
-                        'route' => 'admin/users',
-                        'resource' => Controller\UsersController::class . '::indexAction',
-                    ],
-                ],
-            ],
-            'login' => [
-                'label' => 'Login',
-                'route' => 'brokkr-users:login',
-                'resource' => Controller\AccountController::class . '::loginAction',
-                'class' => 'hollow button',
-            ],
-            'logout' => [
-                'label' => 'Logout',
-                'route' => 'brokkr-users:logout',
-                'resource' => Controller\AccountController::class . '::logoutAction',
-                'class' => 'hollow button',
-            ],
-        ],
-    ],
-    'router' => require __DIR__ . '/module.routes.php',
     'service_manager' => [
         'factories' => [
             AuthenticationService::class => InvokableFactory::class,
