+++ psalm.xml
@@ -8,6 +8,7 @@
     <projectFiles>
         <directory name="src" />
         <ignoreFiles>
+            <directory name="src/Mvc/Test" />
             <directory name="vendor" />
         </ignoreFiles>
     </projectFiles>
