+++ src/Module.php
@@ -2,63 +2,10 @@
 
 namespace Riddlestone\Brokkr\Users;
 
-use Laminas\Authentication\AuthenticationService;
-use Laminas\Mvc\MvcEvent;
-use Laminas\View\Helper\Navigation;
-use Laminas\View\Renderer\PhpRenderer;
-use Riddlestone\Brokkr\Acl\Acl;
-
 class Module
 {
     public function getConfig()
     {
         return require __DIR__ . '/../config/module.config.php';
-    }
-
-    /**
-     * Register {@link onRender} with the application event manager
-     *
-     * @param MvcEvent $event
-     * @return void
-     */
-    public function onBootstrap(MvcEvent $event): void
-    {
-        $event->getApplication()->getEventManager()->attach('render', [$this, 'onRender']);
-    }
-
-    /**
-     * Inject the current authenticated user into {@link Navigation} when rendering
-     *
-     * @param MvcEvent $event
-     * @return void
-     */
-    public function onRender(MvcEvent $event)
-    {
-        $serviceManager = $event->getApplication()->getServiceManager();
-
-        if (!$serviceManager->has('ViewRenderer')) {
-            return;
-        }
-
-        $renderer = $serviceManager->get('ViewRenderer');
-
-        if (!$renderer instanceof PhpRenderer) {
-            return;
-        }
-
-        if (!$renderer->getHelperPluginManager()->has('navigation')) {
-            return;
-        }
-
-        $navigation = $renderer->getHelperPluginManager()->get('navigation');
-
-        if (!$navigation instanceof Navigation) {
-            return;
-        }
-
-        /** @var Navigation $navigation */
-        $navigation
-            ->setAcl($serviceManager->get(Acl::class))
-            ->setRole($serviceManager->get(AuthenticationService::class)->getIdentity());
     }
 }
