+++ src/Mvc/Controller/UsersControllerFactory.php
@@ -1,13 +1,12 @@
 <?php
 
-namespace Riddlestone\Brokkr\Users\Controller;
+namespace Riddlestone\Brokkr\Users\Mvc\Controller;
 
 use Interop\Container\ContainerInterface;
 use Laminas\ServiceManager\Exception\ServiceNotCreatedException;
 use Riddlestone\Brokkr\Users\Repository\UserRepository;
-use Riddlestone\Brokkr\Users\Service\PasswordResetService;
 
-class AccountControllerFactory extends AbstractActionControllerFactory
+class UsersControllerFactory extends AbstractActionControllerFactory
 {
     /**
      * @inheritDoc
@@ -16,17 +15,15 @@ class AccountControllerFactory extends AbstractActionControllerFactory
     {
         $controller = parent::__invoke($container, $requestedName, $options);
 
-        if (!$controller instanceof AccountController) {
+        if (!$controller instanceof UsersController) {
             throw new ServiceNotCreatedException(sprintf(
                 '%s is not an instance of %s',
                 $requestedName,
-                AccountController::class
+                UsersController::class
             ));
         }
 
         $controller->setUserRepository($container->get(UserRepository::class));
-        $controller->setFormElementManager($container->get('FormElementManager'));
-        $controller->setPasswordResetService($container->get(PasswordResetService::class));
 
         return $controller;
     }
