+++ src/Mvc/Test/Integration/config/local.config.php
@@ -1,0 +1,22 @@
+<?php
+
+use Laminas\Router\Http\Literal;
+
+return [
+    'router' => [
+        'routes' => [
+            'home' => [
+                'name' => 'home',
+                'type' => Literal::class,
+                'options' => [
+                    'route' => '/',
+                ],
+            ],
+        ],
+    ],
+    'view_manager' => [
+        'template_path_stack' => [
+            __DIR__ . '/../views',
+        ],
+    ],
+];
