+++ src/Mvc/Test/Unit/Form/LoginFormTest.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace Riddlestone\Brokkr\Users\Test\Unit\Form;
+namespace Riddlestone\Brokkr\Users\Mvc\Test\Unit\Form;
 
 use PHPUnit\Framework\TestCase;
-use Riddlestone\Brokkr\Users\Form\RequestPasswordResetForm;
+use Riddlestone\Brokkr\Users\Mvc\Form\LoginForm;
 
-class RequestPasswordResetFormTest extends TestCase
+class LoginFormTest extends TestCase
 {
     public function testInit()
     {
-        $form = new RequestPasswordResetForm();
+        $form = new LoginForm();
         $form->init();
-        $this->assertCount(2, $form->getElements());
+        $this->assertCount(3, $form->getElements());
     }
 
     public function isValidData()
@@ -24,18 +24,28 @@ class RequestPasswordResetFormTest extends TestCase
             [
                 'data' => [
                     'email_address' => 'someone@example.com',
+                    'password' => 'foobar',
                 ],
                 'valid' => true,
             ],
             [
                 'data' => [
                     'email_address' => '',
+                    'password' => 'foobar',
                 ],
                 'valid' => false,
             ],
             [
                 'data' => [
                     'email_address' => 'email-address-without-at',
+                    'password' => 'foobar',
+                ],
+                'valid' => false,
+            ],
+            [
+                'data' => [
+                    'email_address' => 'someone@example.com',
+                    'password' => '',
                 ],
                 'valid' => false,
             ],
@@ -49,7 +59,7 @@ class RequestPasswordResetFormTest extends TestCase
      */
     public function testIsValid(array $data, bool $valid)
     {
-        $form = new RequestPasswordResetForm();
+        $form = new LoginForm();
         $form->init();
         $form->setData($data);
         $this->assertEquals($valid, $form->isValid());
